+++ src/typescript/applications/actor/wvActorSheet.ts
@@ -7,8 +7,12 @@ import {
   TYPES
 } from "../../constants.js";
 import { getGame } from "../../foundryHelpers.js";
-import { isSkillName, isSpecialName } from "../../helpers.js";
-import { boundsSettingNames } from "../../settings.js";
+import {
+  getSkillMinPoints,
+  getSpecialMinPoints,
+  isSkillName,
+  isSpecialName
+} from "../../helpers.js";
 import WvI18n, { I18nSpecial } from "../../wvI18n.js";
 import { LOG } from "../../systemLogger.js";
 import type { SkillDragData, SpecialDragData } from "../../actor/wvActor.js";
@@ -65,20 +69,8 @@ export default class WvActorSheet extends ActorSheet<
     data.sheet = {};
 
     data.sheet.bounds = CONSTANTS.bounds;
-    const skillMin = getGame().settings.get(
-      CONSTANTS.systemId,
-      boundsSettingNames.skills.points.min
-    );
-    if (typeof skillMin === "number") {
-      data.sheet.bounds.skills.points.min = skillMin;
-    }
-    const specialMin = getGame().settings.get(
-      CONSTANTS.systemId,
-      boundsSettingNames.special.points.min
-    );
-    if (typeof specialMin === "number") {
-      data.sheet.bounds.special.points.min = specialMin;
-    }
+    data.sheet.bounds.skills.points.min = getSkillMinPoints();
+    data.sheet.bounds.special.points.min = getSpecialMinPoints();
 
     const specialI18ns = WvI18n.specials;
     data.sheet.specials = {};